requirements.txt
flask
flask-cors
python-dotenv
mysql-connector-python
aiohttp